Keep Ping application alive without WebFlux
Prior to this commit, the Ping application would have a dependency on
the WebFlux starter to start a server and artificially keep the
application up, even if it doesn't need a server.

Spring Boot applications, in general, are kept up until all
command-line/application runners are done and all non-daemon threads are
stopped.

In this case, if we remove the webflux starter the application has no
non-daemon thread (i.e. no server) and the runners and in this case the
event listener are done quite quickly since the reactive pipelines are
terminated with the `subscribe` operator.
The `subscribe` operator kicks off the processing of the pipeline and
schedules that work on the reactor infrastructure - but nothing waits
for the completion of this work.

This commit adds `.then().block()` operators at the end of those
pipelines to make sure that the application remains up until processing
is done.

Note that the Spring Framework team is currently investigating whether
we can improve the publishing and handling of application events in
reactive applications.